Add classes for Parser, Storage and TaskList
Created a new Storage class which handles
* Reading of data from userdata.txt
* Writing to userdata.txt

Created a new Parser class which handles
* Creating new Tasks (ToDos, Deadlines, EventTask)

Created a new TaskList class which handles
* adding of Task
* removing of Task at index
* marking Task as completed at index
* unmarking Task as incomplete at index
* exporting all the tasks
